adding jwt authentication to memory routes

+++ time-capsule/server/src/routes/memories.ts
@@ -4,8 +4,15 @@ import z from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-	app.get("/memories", async () => {
+	app.addHook("preHandler", async (request) => {
+		await request.jwtVerify();
+	});
+
+	app.get("/memories", async (request) => {
 		const memories = await prisma.memory.findMany({
+			where: {
+				userId: request.user.sub,
+			},
 			orderBy: {
 				createdAt: "asc",
 			},
@@ -19,7 +26,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		});
 	});
 
-	app.get("/memories/:id", async (request) => {
+	app.get("/memories/:id", async (request, response) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -32,6 +39,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
 			},
 		});
 
+		if (!memory.isPublic && memory.userId !== request.user.sub) {
+			return response.status(401).send();
+		}
+
 		return memory;
 	});
 
@@ -42,9 +53,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 			isPublic: z.coerce.boolean().default(false),
 		});
 
-		const { description, image, isPublic } = bodySchema.parse(
-			request.body
-		);
+		const { description, image, isPublic } = bodySchema.parse(request.body);
 
 		const newMemory = await prisma.memory.create({
 			data: {
@@ -58,7 +67,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		return newMemory;
 	});
 
-	app.put("/memories/:id", async (request) => {
+	app.put("/memories/:id", async (request, response) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
@@ -71,11 +80,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
 			isPublic: z.coerce.boolean().default(false),
 		});
 
-		const { description, image, isPublic } = bodySchema.parse(
-			request.body
-		);
+		const { description, image, isPublic } = bodySchema.parse(request.body);
 
-		const memory = await prisma.memory.update({
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return response.status(401).send();
+		}
+
+		memory = await prisma.memory.update({
 			where: {
 				id: id,
 			},
@@ -89,12 +106,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 		return memory;
 	});
 
-	app.delete("/memories/:id", async (request) => {
+	app.delete("/memories/:id", async (request, response) => {
 		const paramsSchema = z.object({
 			id: z.string().uuid(),
 		});
 
 		const { id } = paramsSchema.parse(request.params);
+		
+		let memory = await prisma.memory.findUniqueOrThrow({
+			where: {
+				id,
+			},
+		});
+
+		if (memory.userId !== request.user.sub) {
+			return response.status(401).send();
+		}
 
 		await prisma.memory.delete({
 			where: {
